refactored all the commands into utils file(s)
