+++ src/commands/misc/help.js
@@ -1,4 +1,5 @@
-const {SlashCommandBuilder,  EmbedBuilder } = require('@discordjs/builders');
+const {SlashCommandBuilder} = require('@discordjs/builders');
+const {help} = require('../../utils/miscCommands')
 
 module.exports = {
     data: new SlashCommandBuilder()
@@ -13,26 +14,6 @@ module.exports = {
     },
 
     async execute(client, interaction) {
-        //Create an embed to send to the user
-        const embed = new EmbedBuilder()
-            .setTitle('Server Commands')
-            .addFields(
-                {
-                    // name - /[name of command]
-                    name: '/help-remember',
-                    // value - description of command
-                    value: 'Brings up list of remember commands'
-                },
-                {
-                    name: '/check-in-interface',
-                    value: 'Create an interface for users to schedule their check ins'
-                },
-                {
-                    name: '/schedule-check-in',
-                    value: 'Schedule a day and time to be notified'
-                }
-            );
-
-        interaction.reply({ embeds: [embed], ephemeral: true });
+        interaction.reply(help());
     }
 };
