+++ src/events/messageCreate/checkUnavailability.js
@@ -2,7 +2,6 @@ const { getAvailabilityChannel } = require('../../utils/availability');
 const { getChannelObject } = require('../../utils/apiCalls');
 module.exports = async (client, message) => {
     try {
-
         const availabilityChannel = await getAvailabilityChannel();
         //only send a message if it's not from a bot and it's from the available channel
         if (message.author.bot || !availabilityChannel || message.channelId !== availabilityChannel.id) {
