+++ src/utils/availability.js
@@ -3,7 +3,25 @@ let availabilityChannel = undefined;
 const getAvailabilityChannel = async () => { return availabilityChannel }
 const setAvailabilityChannel = (channel) => { availabilityChannel = channel }
 
+const availability = async (newChannel) => {
+    const oldChannel = await getAvailabilityChannel();
+
+    //check if new id is the same as the current one
+    if (oldChannel && oldChannel.id === newChannel.id) {
+        return{
+            content: `<#${oldChannel.id}> is already the availability channel`
+        }
+    }
+
+    //set the new channel as the current one
+    setAvailabilityChannel(newChannel)
+
+    return{
+        content: `<#${newChannel.id}> is the new availability channel`
+    }
+}
 module.exports = {
     getAvailabilityChannel,
-    setAvailabilityChannel
+    setAvailabilityChannel,
+    availability
 }
